+++ iati/tests/utilities.py
@@ -23,7 +23,7 @@ import iati.tests.resources
 
 
 # This will need updating once test data forcing XML format is fixed
-RULESET_FOR_TESTING = iati.Ruleset(iati.resources.load_as_string(iati.resources.get_ruleset_path('ruleset_for_tests')))
+RULESET_FOR_TESTING = iati.Ruleset(iati.utilities.load_as_string(iati.resources.get_ruleset_path('ruleset_for_tests')))
 """A working Ruleset based on the Standard Ruleset."""
 
 
@@ -34,11 +34,11 @@ SCHEMA_ORGANISATION_NAME_VALID = 'iati-organisations-schema'
 SCHEMA_NAME_VALID = 'iati-activities-schema'
 """A string containing a valid Schema name."""
 
-XML_TREE_VALID = etree.fromstring(iati.tests.resources.load_as_string('valid_not_iati'))
+XML_TREE_VALID = etree.fromstring(iati.utilities.load_as_string(iati.tests.resources.get_test_data_path('valid_not_iati')))
 """An etree that is valid XML but not IATI XML."""
-XML_TREE_VALID_IATI = etree.fromstring(iati.tests.resources.load_as_string('valid_iati'))
+XML_TREE_VALID_IATI = etree.fromstring(iati.utilities.load_as_string(iati.tests.resources.get_test_data_path('valid_iati')))
 """A valid IATI etree."""
-XML_TREE_VALID_IATI_INVALID_CODE = etree.fromstring(iati.tests.resources.load_as_string('valid_iati_invalid_code'))
+XML_TREE_VALID_IATI_INVALID_CODE = etree.fromstring(iati.utilities.load_as_string(iati.tests.resources.get_test_data_path('valid_iati_invalid_code')))
 """A valid IATI etree that has an invalid Code value."""